# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Arabic POS System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: getUsers
      tags: [users]
      summary: Get all users (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UsersListResponse"
    post:
      operationId: createUser
      tags: [users]
      summary: Create user (admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      operationId: getUserById
      tags: [users]
      summary: Get user by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateUser
      tags: [users]
      summary: Update user
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete user (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /sales:
    get:
      operationId: getSales
      tags: [sales]
      summary: Get all sales
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
        - name: search
          in: query
          schema:
            type: string
        - name: startDate
          in: query
          schema:
            type: string
        - name: endDate
          in: query
          schema:
            type: string
        - name: customerId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of sales
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SalesListResponse"
    post:
      operationId: createSale
      tags: [sales]
      summary: Create a new sale
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSaleRequest"
      responses:
        "201":
          description: Sale created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sale"
  /sales/{id}:
    get:
      operationId: getSaleById
      tags: [sales]
      summary: Get sale by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Sale found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sale"
        "404":
          description: Sale not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteSale
      tags: [sales]
      summary: Delete sale
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Sale deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
        errors:
          type: array
          items:
            type: string
      required:
        - message
    SuccessResponse:
      type: object
      properties:
        message:
          type: string
        success:
          type: boolean
      required:
        - message
        - success
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required:
        - total
        - page
        - limit
        - totalPages
...
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, user]
        isActive:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - email
        - role
        - isActive
        - createdAt
        - updatedAt
    CreateUserRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [admin, user]
      required:
        - name
        - email
        - password
        - role
    UpdateUserRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, user]
        isActive:
          type: boolean
        password:
          type: string
    UsersListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/User"
        meta:
          $ref: "#/components/schemas/PaginationMeta"
      required:
        - data
        - meta
...
    SaleItem:
      type: object
      properties:
        productId:
          type: integer
        productName:
          type: string
        quantity:
          type: integer
        unitPrice:
          type: number
        total:
          type: number
      required:
        - productId
        - quantity
        - unitPrice
        - total
    Sale:
      type: object
      properties:
        id:
          type: integer
        invoiceNumber:
          type: string
        customerId:
          type: integer
        customerName:
          type: string
        userId:
          type: integer
        userName:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/SaleItem"
        subtotal:
          type: number
        discount:
          type: number
        tax:
          type: number
        total:
          type: number
        paymentMethod:
          type: string
          enum: [cash, card, bank_transfer]
        notes:
          type: string
        status:
          type: string
          enum: [completed, cancelled, refunded]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - invoiceNumber
        - items
        - subtotal
        - discount
        - tax
        - total
        - paymentMethod
        - status
        - createdAt
        - updatedAt
    CreateSaleRequest:
      type: object
      properties:
        customerId:
          type: integer
        items:
          type: array
          items:
            type: object
            properties:
              productId:
                type: integer
              quantity:
                type: integer
            required:
              - productId
              - quantity
        discount:
          type: number
          default: 0
        tax:
          type: number
          default: 0
        paymentMethod:
          type: string
          enum: [cash, card, bank_transfer]
        notes:
          type: string
      required:
        - items
        - paymentMethod
    SalesListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Sale"
        meta:
          $ref: "#/components/schemas/PaginationMeta"
      required:
        - data
        - meta
...