# lib/api-zod/src/generated/types/usersListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Arabic POS System API
 * OpenAPI spec version: 0.1.0
 */
import type { PaginationMeta } from "./paginationMeta";
import type { User } from "./user";

export interface UsersListResponse {
  data: User[];
  meta: PaginationMeta;
}